requirements.txt
pytest
pycodestyle
pyyaml